+++ src/aiologic/testing/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+
+# SPDX-FileCopyrightText: 2025 Ilya Egorov <0x42005e1f@gmail.com>
+# SPDX-License-Identifier: ISC
+
+from ._executors import (
+    create_executor as create_executor,
+    current_executor as current_executor,
+)